refactor: split test files alongside their source modules
Split queries.test.ts (3106 lines) into 6 test files beside their
source modules in services/queries/:
- metrics.test.ts, daily-summary.test.ts, period-summary.test.ts,
  tags.test.ts, activities.test.ts, productivity.test.ts

Split correlations.test.ts (686 lines) into 5 test files beside
their source modules in services/correlations/:
- baseline.test.ts, hrv-activities.test.ts, activity-impact.test.ts,
  event-probability.test.ts, generic.test.ts

Each test file is self-contained with its own vi.mock() declarations.
